Move DSL headers and rename examples
