+++ CMakeLists.txt
@@ -43,4 +43,4 @@ add_subdirectory(libsyst)
 add_subdirectory(libapp)
 add_subdirectory(libplug)
 add_subdirectory(tests)
-add_subdirectory(examples)
+add_subdirectory(studies)
