+++ studies/study_all.cpp
@@ -1,8 +1,8 @@
-#include <analysis/dsl/Study.h>
-#include <analysis/dsl/Helpers.h>
-#include <analysis/dsl/Display.h>
-#include <analysis/dsl/Plots.h>
-#include <analysis/dsl/Snapshot.h>
+#include <workflow/Study.h>
+#include <workflow/Helpers.h>
+#include <workflow/Display.h>
+#include <workflow/Plots.h>
+#include <workflow/Snapshot.h>
 using namespace analysis::dsl;
 
 int main() {
